% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTime.R
\name{calcTime}
\alias{calcTime}
\title{calcTime computes elapsed time at each interval}
\usage{
calcTime(pcAuxData, what)
}
\arguments{
\item{pcAuxData}{pcAuxData object}

\item{what}{???}
}
\value{
Elapsed time for each imputation step
}
\description{
Returns elapsed time for each imputation step
}
\details{
Returns elapsed time for each imputation step
}
