% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subroutines.R
\name{checkInputs}
\alias{checkInputs}
\title{A collection of helper functions}
\usage{
checkInputs()
}
\description{
Various helper functions used throughout the package
}
\details{
Various helper functions used throughout the package
}
