% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPcAux.R
\name{createPcAux}
\alias{createPcAux}
\title{Create the PcAuxData object, calculate interactions, do PCA}
\usage{
createPcAux(
  pcAuxData,
  nComps,
  interactType = 0L,
  maxPolyPow = 1L,
  simMode = FALSE,
  seed = NULL,
  verbose = 2L,
  doImputation = TRUE,
  castData = !doImputation,
  control,
  micemethods = c("norm", "polr", "polyreg", "logreg"),
  ...
)
}
\arguments{
\item{pcAuxData}{A dataframe cleaned by prepData()}

\item{nComps}{Number of principal components}

\item{interactType}{Choose second order or third order interactions}

\item{maxPolyPow}{???}

\item{simMode}{Use simulation mode}

\item{seed}{Set a seed for reproducible results}

\item{verbose}{Use verbose output}

\item{doImputation}{Do the initial single imputation}

\item{castData}{Cast the data as a specific type}

\item{control}{???}

\item{micemethods}{Imputation methods to pass to mice()}

\item{...}{pass other options to mice()}
}
\description{
Second step of the pcaux imputation process
}
\details{
Second step of the pcaux imputation process
}
