% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_PcAuxData_ExtraMethods.R,
%   R/helperFunctions.R
\name{.onAttach}
\alias{.onAttach}
\title{A collection of helper functions}
\usage{
.onAttach(libname, pkgname)

.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library name}

\item{pkgname}{package name}
}
\description{
Various helper functions used throughout the package

Various helper functions used throughout the package
}
\details{
Various helper functions used throughout the package

Various helper functions used throughout the package
}
