% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImpData.R
\name{getImpData}
\alias{getImpData}
\title{getImpData returns the imputed data sets}
\usage{
getImpData(pcAuxData)
}
\arguments{
\item{pcAuxData}{pcAuxData object}
}
\value{
Imputed data sets. One data set for each nImps.
}
\description{
Returns imputed data sets
}
\details{
Returns imputed data sets
}
