% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{inspect gives S3/S4-like access to fields:}
\usage{
inspect(object, what)
}
\arguments{
\item{object}{Name of object to inspect}

\item{what}{???}
}
\value{
Value of field passed in the object parameter
}
\description{
Returns value of the field passed in object parameter
}
\details{
Returns value of the field passed in object parameter
}
