% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePcAux.R
\name{mergePcAux}
\alias{mergePcAux}
\title{Merge the Principal Component Auxiliaries with the raw data from which they
were created}
\usage{
mergePcAux(pcAuxData, rawData, nComps = NULL, verbose = TRUE, ...)
}
\arguments{
\item{pcAuxData}{pcAuxData object}

\item{rawData}{The raw data to be merged with the principal components}

\item{nComps}{The number of components to merge}

\item{verbose}{Toggle verbosity}

\item{...}{other function parameters}
}
\value{
A raw data set joined with the principal component auxiliary
variables
}
\description{
Merge principal components with raw data
}
\details{
Merge principal components with raw data
}
