% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miWithPcAux.R
\name{miWithPcAux}
\alias{miWithPcAux}
\title{A collection of helper functions}
\usage{
miWithPcAux(
  rawData,
  pcAuxData,
  nImps = 100L,
  nomVars = NULL,
  ordVars = NULL,
  idVars = NULL,
  dropVars = "useExtant",
  nComps = NULL,
  compFormat = "list",
  seed = NULL,
  simMode = FALSE,
  forcePmm = FALSE,
  nProcess = 1L,
  verbose = 2L,
  control,
  micemethods = c("norm", "polr", "polyreg", "logreg")
)
}
\arguments{
\item{rawData}{Raw data}

\item{pcAuxData}{pcAuxData object}

\item{nImps}{Number of imputations to perform}

\item{nomVars}{Nominal Variables}

\item{ordVars}{Ordinal Variables}

\item{idVars}{ID Variables}

\item{dropVars}{Variables dropped from imputation}

\item{nComps}{???}

\item{compFormat}{???}

\item{seed}{Random seed for repeatable results}

\item{simMode}{Use simulation mode}

\item{forcePmm}{Force use of Pmm imputation method}

\item{nProcess}{Number of processors to use}

\item{verbose}{Use verbose mode}

\item{control}{???}

\item{micemethods}{mice::mice() paramenters}
}
\description{
Various helper functions used throughout the package
}
\details{
Various helper functions used throughout the package
}
