% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaux-package.R
\docType{package}
\name{pcaux-package}
\alias{pcaux}
\alias{pcaux-package}
\title{pcaux: Principal Component Auxiliary Variables}
\description{
Generates auxiliary variables.
}
\author{
\strong{Maintainer}: Danny Squire \email{danny.squire@ttu.edu} (\href{https://orcid.org/0000-0001-7423-9274}{ORCID})

}
\keyword{internal}
