% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcAuxL.R
\name{pcAuxL}
\alias{pcAuxL}
\title{pcAuxL prints the GPL-3 that describe the licensing conditions for PcAux.}
\usage{
pcAuxL()
}
\value{
License Information
}
\description{
pcAuxL prints the GPL-3 that describe the licensing conditions for PcAux.
}
