% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedHelperFunctions.R, R/pcAuxW.R
\name{pcAuxW}
\alias{pcAuxW}
\title{Create the PcAuxData object, calculate interactions, do PCA}
\usage{
pcAuxW()

pcAuxW()
}
\value{
Warranty Information
}
\description{
Second step of the pcaux imputation process

Displays the warranty information
}
\details{
Second step of the pcaux imputation process

Displays the warranty information
}
