% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcQuickPred.R
\name{pcQuickPred}
\alias{pcQuickPred}
\title{New Predictor Matrix Method to ensure no PCs below a certain threshhold are
used. Also protects against nPredictors > nObservations iteratively.}
\usage{
pcQuickPred(
  data,
  mincor = 0.1,
  minPredCount = 1,
  nLinear = NULL,
  nNonLinear = NULL
)
}
\arguments{
\item{data}{pcAuxData object}

\item{mincor}{Minimum correlation}

\item{minPredCount}{Minimum number of predictors}

\item{nLinear}{Number of linear}

\item{nNonLinear}{Number of non-linear}
}
\value{
mice predictor matrix
}
\description{
pcQuickPred
}
\details{
pcQuickPred
}
