% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare data frame and convert to PcAuxData object}
\usage{
prepData(
  rawData,
  moderators = NULL,
  nomVars = NULL,
  ordVars = NULL,
  idVars = NULL,
  dropVars = NULL,
  groupVars = NULL,
  simMode = FALSE,
  nProcess = 1L,
  verbose = 2L,
  control,
  ...
)
}
\arguments{
\item{rawData}{data frame}

\item{moderators}{moderators}

\item{nomVars}{nomVars}

\item{ordVars}{ordVars}

\item{idVars}{idVars}

\item{dropVars}{dropVars}

\item{groupVars}{groupVars}

\item{simMode}{simMode}

\item{nProcess}{nProcess}

\item{verbose}{verbose}

\item{control}{control}

\item{...}{other}
}
\description{
\code{prepData} checks the input data from and returns a PcAuxData object.
}
\details{
This is the first step in generating auxillary variables.
}
