% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample1.R
\docType{data}
\name{sample1}
\alias{sample1}
\title{Sample 1}
\format{
\subsection{\code{sample1}}{

A data frame with 500 rows and 19 columns:
\describe{
\item{ID}{Unique ID Column}
\item{grade}{School Grade}
\item{age}{Age of respondent}
\item{male}{Gender of respondent}
\item{gpa}{GPA of respondent}
\item{int1, int2, int3, int4, int5, int6}{Simulated Likert scale data}
\item{ext1, ext2, ext4, ext4, ext5, ext6}{Simulated Likert scale data}
\item{incident}{Simulated boolean data}
\item{Qual}{Simulated text data}
...
}
}
}
\source{
Synthetic data set
}
\usage{
sample1
}
\description{
A simple synthetic data set with missing values for demonstration purposes
only.
}
\keyword{datasets}
