% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStatus.R
\name{writeStatus}
\alias{writeStatus}
\title{writeStatus writes status of imputation run to log file}
\usage{
writeStatus(pcAuxData, outName, what)
}
\arguments{
\item{pcAuxData}{pcAuxData object}

\item{outName}{name of the output file}

\item{what}{the status you want to know}
}
\value{
status
}
\description{
Returns status
}
\details{
Returns status
}
