% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTime.R
\name{calcTime}
\alias{calcTime}
\title{calcTime records the system time for the PcAux functions at set intervals to help determine is-
sues with lengthy runs. calcTime allows the user to extract timing information for specific functions
of PcAux individually.}
\usage{
calcTime(pcAuxData, what)
}
\arguments{
\item{pcAuxData}{A fitted object of class PcAuxData produced as output of prepData, createPcAux,
or miWithPcAux functions.}

\item{what}{A character vector indicating the name of a function for which to extract status
information. abbreviates the functions prepData, createPcAux, and miWithPcAux
as "prep", "create", and "mi", respectively.}
}
\value{
A named vector with an entry per interval.
}
\description{
calcTime records the system time for the PcAux functions at set intervals to help determine is-
sues with lengthy runs. calcTime allows the user to extract timing information for specific functions
of PcAux individually.
}
