% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImpData.R
\name{getImpData}
\alias{getImpData}
\title{getImpData is a simple wrapper function that extracts the completed, multiply imputed data sets from a
fitted PcAuxData object produced by running the miWithPcAux function.}
\usage{
getImpData(pcAuxData)
}
\arguments{
\item{pcAuxData}{A fitted object of class PcAuxData produced as output of the miWithPcAux func-
tion.}
}
\value{
Imputed data sets. One data set for each nImps.
}
\description{
getImpData is a simple wrapper function that extracts the completed, multiply imputed data sets from a
fitted PcAuxData object produced by running the miWithPcAux function.
}
