% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoggedEvents.R
\name{getLoggedEvents}
\alias{getLoggedEvents}
\title{getLoggedEvents retrieves logged events from the imputation run}
\usage{
getLoggedEvents(pcAuxData)
}
\arguments{
\item{pcAuxData}{pcAuxData object}
}
\value{
Logged Events
}
\description{
Returns Logged Events
}
\details{
Returns Logged Events
}
