% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{inspect provides S3/S4-like access to fields of a PcAuxData Reference Class object.}
\usage{
inspect(object, what)
}
\arguments{
\item{object}{An initialized RC object of class PcAuxData.}

\item{what}{A character string naming the field to access in object.}
}
\value{
The current value stored in the what field of object.
}
\description{
inspect provides S3/S4-like access to fields of a PcAuxData Reference Class object.
}
