% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePcAux.R
\name{mergePcAux}
\alias{mergePcAux}
\title{Merge the Principal Component Auxiliaries with the raw data from which they
were created.}
\usage{
mergePcAux(pcAuxData, rawData, nComps = NULL, verbose = TRUE, ...)
}
\arguments{
\item{pcAuxData}{An object of class PcAuxData produced by a run of
createPcAux.}

\item{rawData}{A data frame containing the incomplete data for which to
create the multiple imputations.}

\item{nComps}{A two-element vector giving the number of linear and nonlinear,
respectively, component scores to extract. See the Details section for more
information. When not speciﬁed, all component scores that exist in
pcAuxData are used.}

\item{verbose}{An integer code in 0, 1, 2 deﬁning the verbosity of output
printed to the screen. verbose = 0prints no output; verbose = 1prints all
output except for the mes- sages returned by mice; verbose = 2 prints all
output, including the messages returned by mice. Warnings are always
printed, regardless of the value assigned to verbose. Defaults to verbose =
2.}

\item{...}{Unused}
}
\value{
Raw data with Auxiliary data attached.
}
\description{
Merge the Principal Component Auxiliaries with the raw data from which they
were created.
}
