% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePcAux.R
\name{mergePcAux}
\alias{mergePcAux}
\title{Merge Principal Component Auxiliaries with the raw data from which
they were constructed.}
\usage{
mergePcAux(pcAuxData, rawData, nComps = NULL, verbose = TRUE, ...)
}
\arguments{
\item{pcAuxData}{An object of class PcAuxData produced by a call to createPcAux.}

\item{rawData}{A data frame containing the raw data used to run createPcAux.}

\item{nComps}{A two-element vector giving the number of linear and nonlinear, respectively,
component scores to extract. See the Details section for more information.
When not specified, all component scores that exist in pcAuxData are used.}

\item{verbose}{A logical flag indicating whether verbose output should be printed to stdout.
Defaults to verbose = TRUE.}

\item{...}{Only used when mergePcAux is called from within other PcAux functions.}
}
\value{
A data frame with (a subset of) the principal component auxiliary scores from pcAuxData$pcAux
merged onto the end of the raw data.
}
\description{
Merge Principal Component Auxiliaries with the raw data from which
they were constructed.
}
