% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaux-package.R
\docType{package}
\name{PcAux-package}
\alias{PcAux}
\alias{PcAux-package}
\title{PcAux: Principle Components Auxiliary Variables}
\description{
Automatically extract auxiliary features for simple, principled missing data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Statscamp/PcAux}
  \item Report bugs at \url{https://github.com/Statscamp/PcAux/issues}
}

}
\author{
\strong{Maintainer}: Kyle Lang \email{k.m.lang@uvt.nl}

Authors:
\itemize{
  \item Todd Little \email{todd.little@ttu.edu}
}

Other contributors:
\itemize{
  \item Steven Chesnut [contributor]
  \item Vibhuti Gupta [contributor]
  \item Byungkwan Jung [contributor]
  \item Pavel Panko [contributor]
  \item Luke Waggenspack [contributor]
  \item Danny Squire \email{danny.squire@ttu.edu} (\href{https://orcid.org/0000-0001-7423-9274}{ORCID}) [contributor]
  \item Daniel Bontempo \email{daniel.bontempo@ttu.edu} [contributor]
}

}
\keyword{internal}
