\name{pcAuxW}
\alias{pcAuxW}
\title{
  Print warranty statement for \pkg{PcAux}.
}
\description{
  Print the sections of the GPL-3 that describe the warranty (or
  complete lack thereof) for \pkg{PcAux}.
}
\usage{
pcAuxW()
}
\value{
  Text giving the warranty-specific sections of the GPL-3.
}
\author{
  Kyle M. Lang
}
\examples{
## Check PcAux's warranty:
pcAuxW()
}