% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcAuxW.R
\name{pcAuxW}
\alias{pcAuxW}
\title{pcAuxW prints the sections of the GPL-3 that describe the warranty (or complete lack thereof) for PcAux.}
\usage{
pcAuxW()
}
\value{
Text giving the warranty-specific sections of the GPL-3.
}
\description{
pcAuxW prints the sections of the GPL-3 that describe the warranty (or complete lack thereof) for PcAux.
}
