% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportedHelperFunctions.R
\name{pcAuxW}
\alias{pcAuxW}
\title{Create the PcAuxData object, calculate interactions, do PCA}
\usage{
pcAuxW()
}
\description{
Second step of the pcaux imputation process
}
\details{
Second step of the pcaux imputation process
}
