% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\name{prepData}
\alias{prepData}
\title{Prepare a data frame for conversion into a PcAuxData object}
\usage{
prepData(
  rawData,
  moderators = NULL,
  nomVars = NULL,
  ordVars = NULL,
  idVars = NULL,
  dropVars = NULL,
  groupVars = NULL,
  simMode = FALSE,
  nProcess = 1L,
  verbose = 2L,
  control,
  ...
)
}
\arguments{
\item{rawData}{A data frame or matrix containing raw data}

\item{moderators}{moderators}

\item{nomVars}{A list of column names which contain nominal values}

\item{ordVars}{A list of column names which contain ordinal values}

\item{idVars}{A list of column names which contain unique row
identifiers}

\item{dropVars}{A list of column names to exclude from the
imputation}

\item{groupVars}{A list of column names which contains group}

\item{simMode}{simMode}

\item{nProcess}{nProcess}

\item{verbose}{verbose}

\item{control}{control}

\item{...}{other}
}
\value{
An S4 class of type pcAuxData as defined in 01_PcAuxData.R
}
\description{
\code{prepData} checks the input data and returns a PcAuxData object.
}
\details{
This is the first step in generating auxillary variables.
}
