% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStatus.R
\name{writeStatus}
\alias{writeStatus}
\title{writeStatus provides a text file featuring system and OS information for documentation purposes.}
\usage{
writeStatus(pcAuxData, outName, what)
}
\arguments{
\item{pcAuxData}{A fitted object of class PcAuxData produced as output of prepData, createPcAux,
or miWithPcAux functions.}

\item{outName}{A character vector indicating the name of the file to write the information to,
such as a .txt file.}

\item{what}{A character vector indicating the name of a function for which to extract status
information. abbreviates the functions prepData, createPcAux, and miWithPcAux
as "prep", "create", and "mi", respectively.}
}
\value{
A string object containing the compiled system and OS information.
}
\description{
writeStatus provides a text file featuring system and OS information for documentation purposes.
}
